updated functions and structures, got rid of Blinky related stuff

+++ ptolemy/apps/ptidesBareIron/PtidesOS/boards/ek-lm3s8962/SimplePtidesOS/structures.h
@@ -38,7 +38,7 @@ typedef struct{
 
 typedef struct
 {
-    long timestamp;
+    long long timestamp;
     int microstep;
 } Tag;
